Move service join request schemas
This moves them to same directory as the endpoint they are for, to match
where we put our other schemas.
